# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "API for chess tournament administration system.",
        "contact": {
            "email": "xdudasko@stuba.sk, xdemeterp@stuba.sk"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://mtaa_chess_api.test/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/tournaments": {
            "put": {
                "tags": [
                    "Tournaments"
                ],
                "summary": "Store new tournament",
                "description": "Returns tournament data",
                "operationId": "storeTournament",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "date",
                                    "tempo"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Tournament name"
                                    },
                                    "date": {
                                        "type": "datetime",
                                        "format": "YYYY-MM-DD HH:MM:SS",
                                        "example": "2021-05-06 09:00:00"
                                    },
                                    "tempo": {
                                        "type": "integer",
                                        "example": "3"
                                    },
                                    "tempo_increment": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "rounds": {
                                        "type": "integer",
                                        "example": "15"
                                    },
                                    "description": {
                                        "type": "text",
                                        "example": "Description"
                                    },
                                    "image": {
                                        "type": "bytearray",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tournament"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Invalid request body.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, invalid request body."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tournaments/{id}": {
            "get": {
                "tags": [
                    "Tournaments"
                ],
                "summary": "Get tournament information",
                "description": "Returns tournament data",
                "operationId": "getTournamentById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tournament id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tournament"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tournaments"
                ],
                "summary": "Overrides existing tournament values.",
                "description": "Returns updated tournament data",
                "operationId": "updateTournament",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tournament id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Tournament"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tournament"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "422": {
                        "description": "Invalid request body.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, invalid request body."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tournaments"
                ],
                "summary": "Delete existing tournament",
                "description": "Deletes a record and returns no content.",
                "operationId": "deleteTournament",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tournament id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/tournaments/{hash}": {
            "get": {
                "tags": [
                    "Tournaments"
                ],
                "summary": "Check if the scanned hash is a valid tournament hash.",
                "description": "Returns true if the tournament is active and player can attend it.",
                "operationId": "checkTournamentByHash",
                "parameters": [
                    {
                        "name": "hash",
                        "in": "path",
                        "description": "Tournament hash",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Returns true if the provided tournament hash is valid.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "valid": {
                                            "type": "bool",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Tournament": {
                "title": "Tournament",
                "description": "Tournament model",
                "properties": {
                    "title": {
                        "type": "string",
                        "format": "string",
                        "example": "Tournament name"
                    },
                    "date": {
                        "type": "datetime",
                        "format": "YYYY-MM-DD HH:MM:SS",
                        "example": "2021-05-06 09:00:00"
                    },
                    "tempo": {
                        "type": "integer",
                        "example": "3"
                    },
                    "tempo_increment": {
                        "type": "integer",
                        "example": "2"
                    },
                    "rounds": {
                        "type": "integer",
                        "example": "15"
                    },
                    "description": {
                        "type": "text",
                        "example": "Description"
                    },
                    "image": {
                        "type": "bytearray",
                        "example": ""
                    },
                    "qr_hash": {
                        "type": "string",
                        "example": ""
                    },
                    "created_at": {
                        "type": "timestamp",
                        "example": "2021-05-06 09:00:00"
                    },
                    "updated_at": {
                        "type": "timestamp",
                        "example": "2021-05-06 09:00:00"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Tournament"
                }
            },
...
        }
    },
...
    "security": [
        []
    ]
}